Geocoding with Google Maps
I switched OSM's geocoder (Nominatim) for gmaps' GoogleV3. the number of NAs went from 170 to just 1.

the API key is imported from the utils file, but the utils file itself imports the api from the .env file stored in the machine, so make sure you add the API key to your local .env file with the proper name.